+++ CHANGELOG.md
@@ -13,14 +13,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- 
+-
 
 ### Changed
 
-- 
+-
 
 ### Removed
 
+-
+
+## [0.3.0] - 2024-09-07
+
+### Fixed
+
+-
+
+### Added
+
+-
+
+### Changed
+
+- Add std::mutex::Mutex around internal ```__REGISTRY``` static
+- TraitRegStorage initialization implemented internally, rather than in macro generated code
+
+### Removed
+
+- unsafe around static mut internally
+- Default implementation on TraitRegStorage
 - 'examples' and 'tests' directories from package
 
 ## [0.2.0] - 2024-09-06